+++ Krowi_AchievementFilter/Data/ExportedNextPatchAchievements.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-08-06 10-22-49 ]] --
+-- [[ Exported at 2021-08-10 00-03-19 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --