+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 ## [9.1.0.28.0] - 2021-
 ### Added
 - Shadowlands Covenant Assaults as Event Reminders
+- Battle for Azeroth Faction Assaults and N'Zoth Assaults as Event Reminders
 - Options to set the date and time format of the event reminders
+
+### Changed
+- Moved Battle for Azeroth - Dueler's Guild - Frontline Veteran and - Frontline Slayer to Battle for Azeroth - Zones - Kul Tiras & Zandalar - Player vs. Player
 
 ### Fixed
 - Shadowlands - Zones - Korthia, Shadowlands - Dungeons - Tazavesh, the Veiled Market and Shadowlands - Raids - Sanctum of Domination now all should localize properly
@@ -14,6 +18,7 @@ All notable changes to this project will be documented in this file.
 
 ### TODO
 - If an achievement is expanded in Expansions, selecting the Events tab with a category and achievements does something strange
+- Add the option to show the "Time Left" on events instead of start and endtime
 
 [9.1.0.28.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3412562